+++ saml-identity-provider/src/main/java/se/swedenconnect/spring/saml/idp/Saml2IdentityProviderVersion.java
@@ -23,8 +23,8 @@ package se.swedenconnect.spring.saml.idp;
 public final class Saml2IdentityProviderVersion {
 
   private static final int MAJOR = 2;
-  private static final int MINOR = 3;
-  private static final int PATCH = 14;
+  private static final int MINOR = 4;
+  private static final int PATCH = 0;
 
   /**
    * Global serialization value for SAML Identity Provider classes.