+++ docs/release-notes.md
@@ -4,11 +4,11 @@
 
 ![License](https://img.shields.io/badge/License-Apache%202.0-blue.svg) ![Maven Central](https://img.shields.io/maven-central/v/se.swedenconnect.spring.saml.idp/spring-saml-idp.svg)
 
-### Version 2.3.14
+### Version 2.4.0
 
 Date: 
 
-- Handling of received `RequestedAuthnContext` elements in an authentication request has been improved.
+- Handling of received `RequestedAuthnContext` elements in an authentication request has been improved. See [Handling of Requested Authentication Context Class References](configuration.html#handling-of-requested-authentication-context-class-references).
 
 ### Version 2.3.13
 
@@ -18,7 +18,7 @@ Date: 2025-12-19
   threads tried to access the same key. This has been changed and no synchronization is needed anymore.
   See <https://github.com/swedenconnect/credentials-support/issues/102>.
 
-- When may threads attempted to perform encryption at the same time, for the same SP, threading issues could occur. A
+- When many threads attempted to perform encryption at the same time, for the same SP, threading issues could occur. A
   more stable solution is now in place.
 
 ### Version 2.3.12